Initial commit generated by Claude
Prompt was as follows:
"You're helping me create a new repo in python. To start with, we're
going to create a script that tests me on python packaging semantics. To
do so, please create several subfolders where each subfolder represents
a test case. Inside the test case, there will be several files and/or
folders that represent a mini python project. Each test case should
represent a different file/folder structure and have a different
structure/style of imports between them. Some of the examples should be
invalid. The main script at the top level (outside the test cases)
should be a script that reads the contents of the various files from a
given test case (i.e. shows me the code), and then prompts me to make a
prediction on what will happen when the code gets run (run
successfully/error, etc), and then actually run the code in the test
case to see what happens. Please write the entire project to do this. I
want to be able to quiz myself on the intricacies of python packaging
just by running the main file in this repo"